Split the code for the installation source spoke
Move constants, functions and dialogs of the installation source spoke
to a new file.
