+++ pyanaconda/ui/gui/spokes/installation_source.py
@@ -20,19 +20,16 @@
 import time
 import threading
 import os
-import signal
 import re
 
-from collections import namedtuple
 from urllib.parse import urlsplit
 
-from pyanaconda.core import glib, constants
+from pyanaconda.core import constants
 from pyanaconda.core.constants import PAYLOAD_TYPE_DNF, SOURCE_TYPE_HDD, SOURCE_TYPE_URL, \
     SOURCE_TYPE_CDROM, SOURCE_TYPE_NFS, SOURCE_TYPE_HMC, URL_TYPE_BASEURL, URL_TYPE_MIRRORLIST, \
     URL_TYPE_METALINK, SOURCE_TYPE_CLOSEST_MIRROR, SOURCE_TYPE_CDN
-from pyanaconda.core.process_watchers import PidWatcher
 from pyanaconda.flags import flags
-from pyanaconda.core.i18n import _, N_, CN_
+from pyanaconda.core.i18n import _, CN_
 from pyanaconda.modules.common.structures.payload import RepoConfigurationData
 from pyanaconda.modules.common.constants.services import SUBSCRIPTION
 from pyanaconda.payload.image import find_optical_install_media, find_potential_hdiso_sources, \
@@ -41,13 +38,16 @@ from pyanaconda.core.payload import ProxyString, ProxyStringError, parse_nfs_url
 from pyanaconda.core.util import cmp_obj_attrs, id_generator
 from pyanaconda.ui.communication import hubQ
 from pyanaconda.ui.context import context
+from pyanaconda.ui.gui.spokes.lib.installation_source_helpers import validate_proxy, RepoChecks, \
+    ProxyDialog, MediaCheckDialog, IsoChooser, BASEREPO_SETUP_MESSAGE, PROTOCOL_HTTP, \
+    PROTOCOL_HTTPS, PROTOCOL_FTP, PROTOCOL_NFS, PROTOCOL_FILE, PROTOCOL_MIRROR, REPO_PROTO, \
+    CLICK_FOR_DETAILS
 from pyanaconda.ui.helpers import InputCheck, InputCheckHandler, SourceSwitchHandler
 from pyanaconda.ui.lib.subscription import switch_source
-from pyanaconda.ui.gui import GUIObject
-from pyanaconda.ui.gui.helpers import GUIDialogInputCheckHandler, GUISpokeInputCheckHandler
+from pyanaconda.ui.gui.helpers import GUISpokeInputCheckHandler
 from pyanaconda.ui.gui.spokes import NormalSpoke
 from pyanaconda.ui.categories.software import SoftwareCategory
-from pyanaconda.ui.gui.utils import blockedHandler, fire_gtk_action, find_first_child
+from pyanaconda.ui.gui.utils import blockedHandler, fire_gtk_action
 from pyanaconda.ui.gui.utils import gtk_call_once, really_hide, really_show, fancy_set_sensitive
 from pyanaconda.threading import threadMgr, AnacondaThread
 from pyanaconda.payload import utils as payload_utils
@@ -69,17 +69,6 @@ from gi.repository import Gtk
 
 __all__ = ["SourceSpoke"]
 
-BASEREPO_SETUP_MESSAGE = N_("Setting up installation source...")
-
-# These need to match the IDs in protocolComboBox and repoProtocolComboBox in
-# installation_source.glade.
-PROTOCOL_HTTP = 'http'
-PROTOCOL_HTTPS = 'https'
-PROTOCOL_FTP = 'ftp'
-PROTOCOL_NFS = 'nfs'
-PROTOCOL_FILE = 'file'
-PROTOCOL_MIRROR = 'Closest mirror'
-
 # Repo Store Columns
 REPO_ENABLED_COL = 0
 REPO_NAME_COL = 1
@@ -89,315 +78,13 @@ REPO_OBJ = 2
 MODEL_ROW_VALUE = 0
 MODEL_ROW_NAME = 1
 
-REPO_PROTO = {PROTOCOL_HTTP:  "http://",
-              PROTOCOL_HTTPS: "https://",
-              PROTOCOL_FTP:   "ftp://",
-              PROTOCOL_NFS:   "nfs://",
-              PROTOCOL_FILE:  "file://"
-              }
-
-CLICK_FOR_DETAILS = N_(' <a href="">Click for details.</a>')
-
-
-def _validate_proxy(proxy_string, username_set, password_set):
-    """Validate a proxy string and return an input code usable by InputCheck
-
-       :param str proxy_string: the proxy URL string
-       :param bool username_set: Whether a username has been specified external to the URL
-       :param bool password_set: Whether a password has been speicifed external to the URL
-    """
-    proxy_match = URL_PARSE.match(proxy_string)
-    if not proxy_match:
-        return _("Invalid proxy URL")
-
-    # Ensure the protocol is something that makes sense
-    protocol = proxy_match.group("protocol")
-    if protocol and protocol not in ('http://', 'https://', 'ftp://'):
-        return _("Invalid proxy protocol: %s") % protocol
-
-    # Path and anything after makes no sense for a proxy URL
-    # Allow '/' as a path so you can use http://proxy.example.com:8080/
-    if (proxy_match.group("path") and proxy_match.group("path") != "/") \
-            or proxy_match.group("query") or proxy_match.group("fragment"):
-        return _("Extra characters in proxy URL")
-
-    # Check if if authentication data is both in the URL and specified externally
-    if (proxy_match.group("username") or proxy_match.group("password")) \
-       and (username_set or password_set):
-        return _("Proxy authentication data duplicated")
-
-    return InputCheck.CHECK_OK
-
-
-RepoChecks = namedtuple("RepoChecks", ["name_check", "url_check", "proxy_check"])
-
-
-class ProxyDialog(GUIObject, GUIDialogInputCheckHandler):
-    builderObjects = ["proxyDialog"]
-    mainWidgetName = "proxyDialog"
-    uiFile = "spokes/installation_source.glade"
-
-    def __init__(self, data, proxy_url):
-        GUIObject.__init__(self, data)
-
-        self._proxy_ok_button = self.builder.get_object("proxyOkButton")
-        GUIDialogInputCheckHandler.__init__(self, self._proxy_ok_button)
-
-        self.proxy_url = proxy_url
-        self._proxy_check = self.builder.get_object("enableProxyCheck")
-        self._proxy_info_box = self.builder.get_object("proxyInfoBox")
-        self._auth_check = self.builder.get_object("enableAuthCheck")
-        self._proxy_auth_box = self.builder.get_object("proxyAuthBox")
-
-        self._proxy_url_entry = self.builder.get_object("proxyURLEntry")
-        self._proxy_username_entry = self.builder.get_object("proxyUsernameEntry")
-        self._proxy_password_entry = self.builder.get_object("proxyPasswordEntry")
-
-        self._proxy_validate = self.add_check(self._proxy_url_entry, self._check_proxy_url)
-        self._proxy_validate.update_check_status()
-
-    def _check_proxy_url(self, inputcheck):
-        proxy_string = self.get_input(inputcheck.input_obj)
-
-        # Don't set an error icon on empty input, but still consider it an error
-        if not proxy_string:
-            return InputCheck.CHECK_SILENT
-
-        return _validate_proxy(proxy_string, self._is_username_set(), self._is_password_set())
-
-    def _is_username_set(self):
-        return self._proxy_username_entry.is_sensitive() and self._proxy_username_entry.get_text()
-
-    def _is_password_set(self):
-        return self._proxy_password_entry.is_sensitive() and self._proxy_password_entry.get_text()
-
-    # Update the proxy validation check on username and password changes to catch
-    # changes in duplicated authentication data
-    def on_proxyUsernameEntry_changed(self, entry, user_data=None):
-        self._proxy_validate.update_check_status()
-
-    def on_proxyPasswordEntry_changed(self, entry, user_data=None):
-        self._proxy_validate.update_check_status()
-
-    def on_proxy_enable_toggled(self, button, *args):
-        self._proxy_info_box.set_sensitive(button.get_active())
-
-        if button.get_active():
-            self.set_status(self._proxy_validate)
-        else:
-            self._proxy_ok_button.set_sensitive(True)
-
-    def on_proxy_auth_toggled(self, button, *args):
-        self._proxy_auth_box.set_sensitive(button.get_active())
-        self._proxy_validate.update_check_status()
-
-    def refresh(self):
-        GUIObject.refresh(self)
-
-        if not self.proxy_url:
-            self._proxy_check.set_active(False)
-            self.on_proxy_enable_toggled(self._proxy_check)
-            self._auth_check.set_active(False)
-            self.on_proxy_auth_toggled(self._auth_check)
-            return
-
-        try:
-            proxy = ProxyString(self.proxy_url)
-            if proxy.username:
-                self._proxy_username_entry.set_text(proxy.username)
-            if proxy.password:
-                self._proxy_password_entry.set_text(proxy.password)
-            self._proxy_url_entry.set_text(proxy.noauth_url)
-        except ProxyStringError as e:
-            log.error("Failed to parse proxy for ProxyDialog.refresh %s: %s", self.proxy_url, e)
-            return
-
-        self._proxy_check.set_active(True)
-        self._auth_check.set_active(bool(proxy.username or proxy.password))
-        self.on_proxy_enable_toggled(self._proxy_check)
-        self.on_proxy_auth_toggled(self._auth_check)
-
-    def run(self):
-        while True:
-            response = self.window.run()
-
-            if response == 1:
-                if self.on_ok_clicked():
-                    # Ok clicked with valid input, save the proxy data
-                    if self._proxy_check.get_active():
-                        url = self._proxy_url_entry.get_text()
-
-                        if self._auth_check.get_active():
-                            username = self._proxy_username_entry.get_text()
-                            password = self._proxy_password_entry.get_text()
-                        else:
-                            username = None
-                            password = None
-
-                        proxy = ProxyString(url=url, username=username, password=password)
-                        self.proxy_url = proxy.url
-                    else:
-                        self.proxy_url = ""
-                    break
-                else:
-                    # Ok clicked with invalid input, keep running the dialog
-                    continue
-            else:
-                # Cancel or Esc, just exit
-                break
-
-        self.window.destroy()
-
-
-class MediaCheckDialog(GUIObject):
-    builderObjects = ["mediaCheckDialog"]
-    mainWidgetName = "mediaCheckDialog"
-    uiFile = "spokes/installation_source.glade"
-
-    def __init__(self, data):
-        super().__init__(data)
-        self.progressBar = self.builder.get_object("mediaCheck-progressBar")
-        self._pid = None
-
-    def _check_iso_ends_cb(self, pid, status):
-        verify_label = self.builder.get_object("verifyLabel")
-
-        if os.WIFSIGNALED(status):
-            pass
-        elif status == 0:
-            verify_label.set_text(_("This media is good to install from."))
-        else:
-            verify_label.set_text(_("This media is not good to install from."))
-
-        self.progressBar.set_fraction(1.0)
-        glib.spawn_close_pid(pid)
-        self._pid = None
-
-    def _check_iso_stdout_watcher(self, fd, condition):
-        if condition == glib.IOCondition.HUP:
-            return False
-
-        channel = glib.IOChannel(fd)
-        line = channel.readline().strip()
-
-        if not line.isdigit():
-            return True
-
-        pct = float(line)/100
-        if pct > 1.0:
-            pct = 1.0
-
-        self.progressBar.set_fraction(pct)
-        return True
-
-    def run(self, device_path):
-        (retval, self._pid, _stdin, stdout, _stderr) = \
-            glib.spawn_async_with_pipes(None, ["checkisomd5", "--gauge", device_path], [],
-                                        glib.SpawnFlags.DO_NOT_REAP_CHILD |
-                                        glib.SpawnFlags.SEARCH_PATH,
-                                        None, None)
-        if not retval:
-            return
-
-        # This function waits for checkisomd5 to end and then cleans up after it.
-        PidWatcher().watch_process(self._pid, self._check_iso_ends_cb)
-
-        # This function watches the process's stdout.
-        glib.io_add_watch(stdout,
-                          glib.IOCondition.IN | glib.IOCondition.HUP,
-                          self._check_iso_stdout_watcher)
-
-        self.window.run()
-
-    def on_close(self, *args):
-        if self._pid:
-            os.kill(self._pid, signal.SIGKILL)
-
-        self.window.destroy()
-
-
-# This class is responsible for popping up the dialog that allows the user to
-# choose the ISO image they want to use.  We can get away with this instead of
-# selecting a directory because we no longer support split media.
-#
-# Two assumptions about the use of this class:
-# (1) This class is responsible for mounting and unmounting the partition
-#     containing the ISO images.
-# (2) When you call refresh() with a currentFile argument or when you get a
-#     result from run(), the file path you use is relative to the root of the
-#     mounted partition.  In other words, it will not contain the
-#     "/mnt/isodir/install" part.  This is consistent with the rest of anaconda.
-class IsoChooser(GUIObject):
-    builderObjects = ["isoChooserDialog", "isoFilter"]
-    mainWidgetName = "isoChooserDialog"
-    uiFile = "spokes/installation_source.glade"
-
-    def __init__(self, data):
-        super().__init__(data)
-        self._chooser = self.builder.get_object("isoChooserDialog")
-
-        # Hide the places sidebar, since it makes no sense in this context
-        # This is discouraged, but the alternative suggested is to reinvent the
-        # wheel. See also https://bugzilla.gnome.org/show_bug.cgi?id=751730
-        places_sidebar = find_first_child(self._chooser,
-                                          lambda x: isinstance(x, Gtk.PlacesSidebar))
-        if places_sidebar:
-            really_hide(places_sidebar)
-
-    # pylint: disable=arguments-differ
-    def refresh(self, currentFile=""):
-        super().refresh()
-        self._chooser.connect("current-folder-changed", self.on_folder_changed)
-        self._chooser.set_filename(constants.ISO_DIR + "/" + currentFile)
-
-    def run(self, device_name):
-        retval = None
-        device_path = payload_utils.get_device_path(device_name)
-
-        # FIXME: Use a unique mount point.
-        mounts = payload_utils.get_mount_paths(device_path)
-        mountpoint = None
-        # We have to check both ISO_DIR and the DRACUT_ISODIR because we
-        # still reference both, even though /mnt/install is a symlink to
-        # /run/install.  Finding mount points doesn't handle the symlink
-        if constants.ISO_DIR not in mounts and constants.DRACUT_ISODIR not in mounts:
-            # We're not mounted to either location, so do the mount
-            mountpoint = constants.ISO_DIR
-            payload_utils.mount_device(device_name, mountpoint)
-
-        # If any directory was chosen, return that.  Otherwise, return None.
-        rc = self.window.run()
-        if rc == Gtk.ResponseType.OK:
-            f = self._chooser.get_filename()
-            if f:
-                retval = f.replace(constants.ISO_DIR, "")
-
-        if not mounts:
-            payload_utils.unmount_device(device_name, mountpoint)
-
-        self.window.destroy()
-        return retval
-
-    # There doesn't appear to be any way to restrict a GtkFileChooser to a
-    # given directory (see https://bugzilla.gnome.org/show_bug.cgi?id=155729)
-    # so we'll just have to fake it by setting you back to inside the directory
-    # should you change out of it.
-    def on_folder_changed(self, chooser):
-        d = chooser.get_current_folder()
-        if not d:
-            return
-
-        if not d.startswith(constants.ISO_DIR):
-            chooser.set_current_folder(constants.ISO_DIR)
-
 
 class SourceSpoke(NormalSpoke, GUISpokeInputCheckHandler, SourceSwitchHandler):
     """
        .. inheritance-diagram:: SourceSpoke
           :parts: 3
     """
-    builderObjects = ["isoChooser", "isoFilter", "partitionStore", "sourceWindow",
-                      "dirImage", "repoStore"]
+    builderObjects = ["partitionStore", "sourceWindow", "dirImage", "repoStore"]
     mainWidgetName = "sourceWindow"
     uiFile = "spokes/installation_source.glade"
     category = SoftwareCategory
@@ -1273,7 +960,7 @@ class SourceSpoke(NormalSpoke, GUISpokeInputCheckHandler, SourceSwitchHandler):
         if not (repo.proxy or proxy_obj.username or proxy_obj.password):
             return InputCheck.CHECK_OK
 
-        return _validate_proxy(proxy_obj.noauth_url, proxy_obj.username, proxy_obj.password)
+        return validate_proxy(proxy_obj.noauth_url, proxy_obj.username, proxy_obj.password)
 
     # Signal handlers.
     def on_source_toggled(self, button, relatedBox):
